+++ ChangeLog.rst
@@ -7,5 +7,6 @@
 
 * First version. Contains following matchers:
   - ``(has-alist-entries :foo 1 :bar 2)``;
+  - ``(has-plist-entries :foo 1 :bar 2)``;
   - ``(any)`` or just ``_``.
   - ``(contains item1 item2 ...)`` and ``(contains-in-any-order ...)``